Add Raspberry Pi test results
Add output from running the benchmark on a Raspberry Pi. This shows
the RPi-optimized armv5te memcpy to be slightly superior to the
standard memcpy (libcofi), with significant gains for the
randomly aligned case at virtually all sizes. When the size varies
randomly from 1 to 64 or 1 to 1024, standard memcpy is slightly
faster. For the aligned case, there are some small differences,
for example standard memcpy is about 2% faster for page aligned
copy of size 1024, but these differences could depend on
shortcoming of tbe benchmark's memory address selection.