Add ARM_ALIGN option to optimized assembler memcpy
When ARM_ALIGN is defined as 32, memcpy_armv5te_no_overfetch will
use 32-byte aligned block writes instead of 16-byte aligned block
writes (a block write is always 32 bytes). At the moment this
only applies when the source and destination have the same
alignment within a word.

32-byte aligned block writes improves performance for copying
regions of memory that are word-aligned case and may be important
for normal memcpy usage.

Note that the current version of memcpy_armv5te_no_overfetch has
a performance regression for sizes < 36 bytes when ARM_ALIGN is
defined as 32.