Add non-prefetching and early prefetching variants
Add memcpy variants with early prefetch and one that does no prefetch
at all. On the Allwinner platform, no prefetch at all seems to be
fastest, suggesting that the "pld" instruction actually does
nothing at all on this platform.

Also fix a performance bug in the non-overfetching variants that
resulted in overfetching beyond the source region.