Enable PRELOAD_CATCH_UP loop for unaligned case
Re-enable an optimized loop from a previous version to catch up
preloads for the unaligned case. The catching up happens while
blocks are being processed.

PRELOAD_CATCH_UP now clearly shows an improvement for unaligned
copies in DRAM (random size up to 1024 bytes). The pixman style
libarmmem benchmark does not show this improvement, maybe a slight
decrease in the memory benchmark ("M"), but this is probably due
to the way the size is selected (fixed size of 1MB - 64).