Optimize small sizes for unaligned case
Implement a special path for sizes < 8 for the unaligned case.

Use a small size threshold of 68 by default for write_align of 32
of 16; only if the source and destination have a different alignment
within a word, use a small size threshold of 36 (in this case, the
code path for larger sizes is faster).