Implement write alignment of 64, add macro parameters, change benchmarks
Implement write alignment of 64 in the memcpy_simple variants.

The CHECK_SMALL_SIZE_ALIGNMENT and PRELOAD_CATCH_UP defines are converted
to macro parameters for the memcpy_simple variants.

Change the benchmark program with the first three tests trying to more
closely mimic real-world memcpy usage. The tests are aligned copies with
size of a power of two up to 4096, aligned copies with size of a multiple
of four up to 1024, and unaligned copies of sizes up to 1023.