Optimize new memcpy for RPi, and apply unaligned write alignment
Apply write alignment for the unaligned case.

Optimize for RPi, write alignment of 8 bytes seems to work best. Note
however that the function still performs worse than libcofi in most
tests.

Apply write alignment of 32 bytes to RPi memset instead of 8 bytes,
because it seems to faster in this implementation.