Optimize NEON based memcpy, and default to preload line size of 32
Optimize NEON-based memcpy with fast path for sizes <= 256 bytes.
Also change the placing of the preload instruction in the aligned
case to increase performance for large DRAM copies.

Make a preload line size of 32 bytes the default to reflect the
fact that even when the physical cache line size is 64, 32-byte
stride preloads may be needed to properly preload from DRAM.